Added description
This code is a basic introduction to representing the Legendre Equation as numerical approximation done using python. In addition to the derivative approximation of the Legendre equation I go on to find the eigenvectors and eigenvalues of the Legendre operator using the Gram-Schmidt procedure. All calculations are represented visually acting upon a variety of functions via plotting through matplotlib.